Update SP - 2190830

+++ _High_Level_Driver/string_advance.h
@@ -24,7 +24,7 @@ typedef enum
     BASE_16 = 16
 } STR_BASE_t;
 
-void transform_uint8_t_tab_to_string(char *p_buffer, uint16_t buffer_length, uint8_t *data, uint8_t length, STR_BASE_t _base);
+void transform_uint8_t_tab_to_string(char *p_buffer, uint16_t buffer_length, uint8_t *data, uint16_t length, STR_BASE_t _base);
 char *str_tolower (const char *ct);
 char *str_toupper (const char *ct);
 int str_istr (const char *cs, const char *ct);
